Fixed two missing java interfaces
git-svn-id: file:///home/roberto/Git/svn@8494 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 1fb881cfffd996dae33c9cbe299eba0c18634e4a

+++ trunk/xmipp/libraries/bindings/java/xmipp_ImageDouble.cpp
@@ -60,7 +60,7 @@ JNIEXPORT void JNICALL Java_xmipp_ImageDouble_readApplyGeo(
 	if (image != NULL) {
 		if (metadata != NULL) {
 			try {
-				image->readApplyGeo(*metadata, (long) id);
+				image->readApplyGeo(*metadata, (size_t) id);
 			} catch (XmippError xe) {
 				msg = xe.getDefaultMessage();
 			} catch (std::exception& e) {
